Move api helpers to own file
