+++ src/javascript/lib/ticket_statuses.js
@@ -1,0 +1,1 @@
+export default ['new', 'open', 'solved', 'pending', 'hold', 'closed']
